refactored network utilities for usage

+++ include/gui.h
@@ -15,7 +15,7 @@ typedef struct
     uint32_t id;
     uint32_t x;
     uint32_t y;
-    uint16_t direction;
+    int      direction;
 } player;
 
 void initialize_gui(const game *g, const player *p, const player *p2);
